# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Reserva Espacios",
        "description": "API para gestionar reservas de espacios",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/espacios": {
            "get": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Obtener todos los espacios",
                "operationId": "5b30a51c97d21af853cacca0c49ab978",
                "responses": {
                    "200": {
                        "description": "Lista de espacios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Espacio"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener los espacios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al obtener los espacios."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Crear un nuevo espacio",
                "operationId": "d03ef5ee956c549c1a8371ae4d24fc39",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "descripcion",
                                    "capacidad"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Sala de Conferencias"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Una sala amplia para conferencias."
                                    },
                                    "capacidad": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Espacio creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Espacio"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear el espacio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al crear el espacio."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/espacios/{id}": {
            "get": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Obtener un espacio por ID",
                "operationId": "e426367fc6788a5cc54af7230faa86ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del espacio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Espacio"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Espacio no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Espacio no encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener el espacio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al obtener el espacio."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Actualizar un espacio por ID",
                "operationId": "3ada5c4d5167b627fced5871ab47aa45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Sala de Reuniones"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Sala equipada para reuniones."
                                    },
                                    "capacidad": {
                                        "type": "integer",
                                        "example": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Espacio actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Espacio"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Espacio no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Espacio no encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar el espacio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al actualizar el espacio."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Eliminar un espacio por ID",
                "operationId": "4cfb60fb59c3e4c3758125082782c32d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Espacio eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Espacio eliminado correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Espacio no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Espacio no encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al eliminar el espacio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al eliminar el espacio."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Espacio": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Sala de Conferencias"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Una sala amplia para conferencias."
                    },
                    "capacidad": {
                        "type": "integer",
                        "example": 50
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT token in the format **Bearer &lt;token&gt;**"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}